requirements_test.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
pytest-cov>=4.0.0
bleak>=0.20.0
